Bucket data placement form refactoring
