+++ frontend/src/app/components/bucket/bucket-data-placement-form/bucket-data-placement-form.html
@@ -22,7 +22,6 @@
             </div>
         </div>
 
-
         <button class="btn push-next" ko.click="onEditBucketQuota">
             Edit Bucket Quota
         </button>
@@ -31,20 +30,9 @@
         </button>
     </div>
 
-    <data-table class="placement-table" params="
-        columns: placementTableColumns,
-        data: nodePools,
-        rowFactory: pool => createPlacementRow(pool),
-        emptyMessage: 'No resources participate in policy'
-    ">
-        <template name="custom-link">
-            <!-- ko if: $data().href -->
-            <a class="link" data-bind="text: $data().text, href: $data().href"></a>
-            <!-- /ko -->
-            <!-- ko ifnot: $data().href -->
-            <span data-bind="text: $data().text"></span>
-            <!-- /ko -->
-        </template>
+    <bucket-data-placement-table class="push-next" params="nodePools: nodePools">
+    </bucket-data-placement-table>
 
-    </data-table>
+    <bucket-spillover>
+    </bucket-spillover>
 </section>
