+++ frontend/src/app/components/bucket/bucket-data-placement-form/bucket-data-placement-form.js
@@ -2,35 +2,10 @@
 
 import template from './bucket-data-placement-form.html';
 import BaseViewModel from 'components/base-view-model';
-import PlacementRowViewModel from './placement-row';
 import ko from 'knockout';
 import { systemInfo } from 'model';
 import { deepFreeze } from 'utils/core-utils';
 import { openEditBucketQuotaModal, openBucketPlacementPolicyModal } from 'dispatchers';
-
-const placementTableColumns = deepFreeze([
-    {
-        name: 'state',
-        type: 'icon'
-    },
-    {
-        name: 'type',
-        type: 'icon'
-    },
-    {
-        name: 'resourceName',
-        type: 'custom-link'
-    },
-    {
-        name: 'onlineNodeCount',
-        label: 'online nodes in pool'
-    },
-    {
-        name: 'capacity',
-        label: 'Resource Capacity',
-        type: 'capacity'
-    }
-]);
 
 const placementTypeMapping = deepFreeze({
     SPREAD: 'Spread',
@@ -40,8 +15,6 @@ const placementTypeMapping = deepFreeze({
 class BucketDataPlacementFormViewModel extends BaseViewModel {
     constructor({ bucket }) {
         super();
-
-        this.placementTableColumns = placementTableColumns;
 
         this.bucketName = ko.pureComputed(
             () => ko.unwrap(bucket) && ko.unwrap(bucket).name
@@ -85,10 +58,6 @@ class BucketDataPlacementFormViewModel extends BaseViewModel {
                 pool => pool.resource_type === 'CLOUD'
             ).length
         );
-    }
-
-    createPlacementRow(pool) {
-        return new PlacementRowViewModel(pool);
     }
 
     onEditBucketQuota() {
