+++ frontend/src/app/components/bucket/bucket-data-placement-form/bucket-data-placement-form.less
@@ -7,27 +7,4 @@ bucket-data-placement-form  {
     section + section {
         margin: @gutter * 3 0;
     }
-
-    .placement-table {
-        .state-col {
-            min-width: 60px;
-        }
-
-        .type-col {
-            min-width: 60px;
-        }
-
-        .resource-name-col {
-            text-align: left;
-            width: 100%;
-        }
-
-        .online-node-count-col {
-            min-width: 160px;
-        }
-
-        .capacity-col {
-            min-width: 120px;
-        }
-    }
 }
