+++ frontend/src/app/components/bucket/bucket-data-placement-table/bucket-data-placement-table.less
@@ -1,0 +1,28 @@
+/* Copyright (C) 2016 NooBaa */
+
+bucket-data-placement-table {
+    display: block;
+
+    .placement-table {
+        .state-col {
+            min-width: 60px;
+        }
+
+        .type-col {
+            min-width: 60px;
+        }
+
+        .resource-name-col {
+            text-align: left;
+            width: 100%;
+        }
+
+        .online-node-count-col {
+            min-width: 160px;
+        }
+
+        .capacity-col {
+            min-width: 120px;
+        }
+    }
+}
