+++ frontend/src/app/components/bucket/bucket-data-placement-table/placement-row.js
