+++ frontend/src/app/components/bucket/bucket-panel/bucket-panel.html
@@ -54,6 +54,4 @@
         </bucket-cloud-sync-form>
     </div>
 </div>
-
-<bucket-spillover class="greedy"></bucket-spillover>
 <!-- /ko -->
